requirements.txt
python-telegram-bot==21.6
requests==2.32.3
hermes-agent @ git+https://github.com/NousResearch/hermes-agent.git
